% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumberize.R
\name{plumberize}
\alias{plumberize}
\title{Plumberize}
\usage{
plumberize(package_name, out_file = NULL)
}
\arguments{
\item{package_name}{A string with the name of the package for which to generate the API.
The package should be installed. E.g., `plumberize("chatgpt")`.}

\item{out_file}{The name of the file where to save the `plumber.R` code. If `NULL` it will
return it as a string.}
}
\description{
Generate the `plumber.R` file for any package. The `plumber.R` file is the configuration file
that builds the API by using the Plumber framework.
This function will create one endpoint per exported function in the `package_name` package.
}
