% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumberize_function.R
\name{plumberize_function}
\alias{plumberize_function}
\title{Plumberize a Function}
\usage{
plumberize_function(function_name)
}
\arguments{
\item{function_name}{The name of the function to plumberize.}
}
\description{
Plumberize a Function
}
